Add files via upload
so, I did upload the java code for an ohm's law calculator before, but this is the python code for the same calculator. 
I uploaded this so that users can get a kind of comparison between java and python codes.
something I understood was that java is best for beginners but sometimes python is more time efficient cause it uses less lines for codes than in java.