Add files via upload
So, the first code is an ohm's law calculator coded using java. The second one is same the first but coded in python. Users may get an understanding of the difference between the 2 languages. as to my point of view I think that java is good for beginners and python consumes less coding time due to the less numbers of lines in the code.